backend/src/Detections/requirements.txt
opencv-python
ultralytics
numpy
aiohttp
python-dotenv
requests
asyncio
pillow